requirements.txt
streamlit
openai>=1.0.0
requests
python-dotenv